v0.2.0: skeg-telemetry dynamic registry + server bump
Add a downstream-extensible metric registry to skeg-telemetry so
consumer crates (skeg-kv-cache, skeg-tenant, applications on top of
skeg) can register their own counters, histograms, and gauges without
patching the engine's closed enums.

- register_counter / register_histogram / register_gauge return
  &'static handles into fixed-size static pools (256 counter slots,
  64 histogram slots, 64 gauge slots).
- register_op!("base") macro derives <base>_total (sharded counter)
  and <base>_duration_seconds (histogram) at compile time.
- Pool registration is idempotent: repeated calls with the same name
  return the same &'static handle, so the OnceLock::get_or_init
  pattern works without a central init function.
- Hot-path cost preserved: ~4.7 ns for the closed-enum record_op,
  ~2 ns for a dynamic handle tick after OnceLock warmup. The 50 ns
  criterion gate continues to pass.
- Histogram bucket count bumped from 22 to 26 (1 µs to 16.78 s) so
  downstream consumers with longer-tail ops do not clip into +Inf.
- SKEG.STATS and /metrics now append the dynamic registry (sorted by
  name) after the engine's static block; engine schema is grep-stable.

skeg-server bumps to 0.2.0 to track the new telemetry dep surface and
keep the user-facing release tag (Homebrew, GitHub Release) aligned
with skeg --version. skeg-core stays at 0.1.2 on crates.io; its local
dep req is bumped to skeg-telemetry = "0.2" for workspace builds and
will carry out on its next release.

Refactor the static array initialisers in skeg-telemetry/src/metrics.rs
to use inline [const { ... }; N] blocks so clippy stops flagging the
named "const Z" pattern.

+++ crates/skeg-server/src/shard.rs
@@ -465,8 +465,6 @@ fn is_mutation(req: &ShardReq) -> bool {
     )
 }
 
-#[allow(clippy::too_many_lines)]
-// one arm per shard request kind; splitting hurts readability
 // TODO(telemetry): wire `Gauge::VindexVectors` and `Gauge::VindexSizeBytes`
 // at the four vindex-mutating arms below (`VindexCreate`, `VindexDrop`,
 // `Vset`, `Vdel`). Today they read `0` from STATS/metrics. The wiring
@@ -703,7 +701,7 @@ async fn process(
             // Side-effect: refresh telemetry gauges from live vlog counters.
             // These are gauges (current values, not monotonic counters), so
             // polling them via STATS does not double-count anything.
-            skeg_telemetry::set_gauge(skeg_telemetry::Gauge::VlogLiveBytes, bytes as u64);
+            skeg_telemetry::set_gauge(skeg_telemetry::Gauge::VlogLiveBytes, bytes);
             ShardResp::Stats(bytes, evictions, n_keys, budget)
         }
     }
